Move kernel source code to a kernel folder
And remove the include folder. Just move includes
into the src folder next to their .c files. It was a big
PITA to mentally grep to find them in the file list.
This makes it much easier. They're not publically
imported headers anyway, so no need to have them
in a separate folder.